Switch numeric tags to double-precision floating point; Change cg-conv to not multiply weights by default; Detect endianness at build time
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@12190 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/Tag.hpp
@@ -81,7 +81,7 @@ public:
 	static UFILE *dump_hashes_out;
 
 	C_OPS comparison_op;
-	int32_t comparison_val;
+	double comparison_val;
 	uint32_t type;
 	uint32_t comparison_hash;
 	uint32_t dep_self, dep_parent;
